Refactored `CoursesStudentPage` to `CourseDashboard` and updated references for consistency; reorganized `CourseCreationProcessModal` under the `course-editor` folder.
