+++ frontend/src/features/courses/course-editor/course-creation-process-modal.tsx
